requirements.txt
yt-dlp>=2025.12.8
ffmpeg-python>=0.2.0
